# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: 뭉치장 API
  description: |
    유행 음식 공동구매 플랫폼 뭉치장의 API 명세입니다.

    ## 인증
    로그인이 필요한 API는 `Authorization: Bearer {accessToken}` 헤더를 포함해야 합니다.

    ## 공통 응답 형식
    모든 응답은 아래 구조를 따릅니다.
    ```json
    { "success": true, "data": { ... }, "error": null }
    { "success": false, "data": null, "error": { "code": "ERROR_CODE", "message": "설명", "detail": "상세" } }
    ```
  version: v1
servers:
  - url: https://api.moongchijang.com
    description: 운영 서버
  - url: https://api.moongchijang.com/dev
    description: 개발 서버
  - url: http://localhost:8080
    description: 로컬 서버
...
paths:
...
  /api/v1/owner/settlements/monthly-summary:
    get:
      tags: [Owner]
      summary: 사장님 월별 정산 예정 조회
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2026
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 5
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOwnerSettlementMonthlySummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/owner/settlements/month-chips:
    get:
      tags: [Owner]
      summary: 사장님 정산 월 칩 조회
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOwnerSettlementMonthChipList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/owner/settlements/items:
    get:
      tags: [Owner]
      summary: 사장님 정산 공구 카드 목록 조회
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            minimum: 2000
            maximum: 2100
            example: 2026
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 5
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOwnerSettlementItemList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/owner/settlements/refund-requests:
    get:
      tags: [Owner]
      summary: 사장님 환불 요청 목록 조회
      security:
        - bearerAuth: []
      parameters:
        - name: tab
          in: query
          required: false
          schema:
            type: string
            enum: [ALL, PENDING, COMPLETED]
            default: ALL
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOwnerRefundRequestList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/owner/settlements/refund-requests/{participationId}:
    get:
      tags: [Owner]
      summary: 사장님 환불 요청 상세 조회
      security:
        - bearerAuth: []
      parameters:
        - name: participationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOwnerRefundRequestDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/owner/settlements/refund-requests/{participationId}/review-submissions:
    post:
      tags: [Owner]
      summary: 사장님 환불 요청 검토 제출
      security:
        - bearerAuth: []
      parameters:
        - name: participationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerRefundReviewSubmitRequest'
      responses:
        '200':
          description: 제출 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOwnerRefundReviewSubmit'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
    Forbidden:
      description: 권한 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
    Conflict:
      description: 상태 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
...
  schemas:
...
    ApiResponseError:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: false }
        data: { nullable: true, example: null }
        error:
          type: object
          required: [code, message, detail]
          properties:
            code: { type: string, example: GROUPBUY_NOT_FOUND }
            message: { type: string, example: 존재하지 않는 공구입니다 }
            detail: { type: string, nullable: true }

    # ── Auth ───────────────────────────────────────────────────
...
    ApiResponseOwnerSettlementMonthlySummary:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/OwnerSettlementMonthlySummary'
        error: { nullable: true, example: null }
    OwnerSettlementMonthlySummary:
      type: object
      required: [year, month, settlementExpectedAmount, grossRevenueAmount, refundFeeAmount]
      properties:
        year: { type: integer, example: 2026 }
        month: { type: integer, example: 5 }
        settlementExpectedAmount: { type: integer, format: int64, example: 216000 }
        grossRevenueAmount: { type: integer, format: int64, example: 240000 }
        refundFeeAmount: { type: integer, format: int64, example: 24000 }
    ApiResponseOwnerSettlementMonthChipList:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/OwnerSettlementMonthChipList'
        error: { nullable: true, example: null }
    OwnerSettlementMonthChipList:
      type: object
      required: [chips]
      properties:
        chips:
          type: array
          items:
            $ref: '#/components/schemas/OwnerSettlementMonthChip'
    OwnerSettlementMonthChip:
      type: object
      required: [year, month, label]
      properties:
        year: { type: integer, example: 2026 }
        month: { type: integer, example: 5 }
        label: { type: string, example: "2026년 5월" }
    ApiResponseOwnerSettlementItemList:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/OwnerSettlementItemList'
        error: { nullable: true, example: null }
    OwnerSettlementItemList:
      type: object
      required: [year, month, items]
      properties:
        year: { type: integer, example: 2026 }
        month: { type: integer, example: 5 }
        items:
          type: array
          items:
            $ref: '#/components/schemas/OwnerSettlementItem'
    OwnerSettlementItem:
      type: object
      required: [groupBuyId, productName, participantCount, pickupDate, amount, settlementStatus]
      properties:
        groupBuyId: { type: integer, format: int64, example: 901001 }
        productName: { type: string, example: "딸기 타르트" }
        participantCount: { type: integer, example: 18 }
        pickupDate: { type: string, format: date, example: "2026-05-02" }
        amount: { type: integer, format: int64, example: 432000 }
        settlementStatus:
          $ref: '#/components/schemas/OwnerSettlementStatus'
    OwnerSettlementStatus:
      type: string
      description: 사장님 공구 정산 상태
      enum: [SETTLEMENT_COMPLETED, SETTLEMENT_PENDING, REFUND_PROCESSING]
    ApiResponseOwnerRefundRequestList:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/OwnerRefundRequestList'
        error: { nullable: true, example: null }
    OwnerRefundRequestList:
      type: object
      required: [pendingCount, completedCount, hasPendingItems, items]
      properties:
        pendingCount: { type: integer, example: 1 }
        completedCount: { type: integer, example: 2 }
        hasPendingItems: { type: boolean, example: true }
        items:
          type: array
          items:
            $ref: '#/components/schemas/OwnerRefundRequestListItem'
    OwnerRefundRequestListItem:
      type: object
      required: [participationId, groupBuyId, productName, paymentAmount, requesterName, requesterCode, refundReasonLabel, requestedDate, status, exceeded24Hours]
      properties:
        participationId: { type: integer, format: int64, example: 1001 }
        groupBuyId: { type: integer, format: int64, example: 901001 }
        productName: { type: string, example: "초코 크루아상&소금빵 세트" }
        paymentAmount: { type: integer, example: 24000 }
        requesterName: { type: string, example: "김민지" }
        requesterCode: { type: string, example: "P001" }
        refundReasonLabel: { type: string, example: "픽업 불가" }
        requestedDate: { type: string, format: date, example: "2026-05-02" }
        status:
          type: string
          enum: [PENDING, COMPLETED]
        exceeded24Hours: { type: boolean, example: true }
    ApiResponseOwnerRefundRequestDetail:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/OwnerRefundRequestDetail'
        error: { nullable: true, example: null }
    OwnerRefundRequestDetail:
      type: object
      required: [participationId, groupBuyId, productName, requesterName, requestedDate, paymentAmount, penaltyAmount, refundExpectedAmount, status]
      properties:
        participationId: { type: integer, format: int64, example: 1001 }
        groupBuyId: { type: integer, format: int64, example: 901001 }
        productName: { type: string, example: "초코 크루아상&소금빵 세트" }
        requesterName: { type: string, example: "김민지" }
        requestedDate: { type: string, format: date, example: "2026-04-29" }
        paymentAmount: { type: integer, example: 24000 }
        penaltyAmount: { type: integer, example: 2400 }
        refundExpectedAmount: { type: integer, example: 21600 }
        refundReasonDetail:
          type: string
          nullable: true
          example: "일정 변경으로 픽업이 어려워졌어요. 환불 부탁드립니다"
        status:
          type: string
          enum: [PENDING, COMPLETED]
    OwnerRefundReviewActionType:
      type: string
      enum: [APPROVE, DISPUTE]
    OwnerRefundReviewSubmitRequest:
      type: object
      required: [action]
      properties:
        action:
          $ref: '#/components/schemas/OwnerRefundReviewActionType'
        disputeReason:
          type: string
          nullable: true
          maxLength: 500
          description: action이 DISPUTE일 때 입력
          example: "픽업 요청 시간 변경이 선행되어야 합니다."
    ApiResponseOwnerRefundReviewSubmit:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/OwnerRefundReviewSubmitResponse'
        error: { nullable: true, example: null }
    OwnerRefundReviewSubmitResponse:
      type: object
      required: [participationId, processed]
      properties:
        participationId: { type: integer, format: int64, example: 1001 }
        processed: { type: boolean, example: true }
...